Cleanup GL helpers and debug logging
