+++ libraries/gl/src/gl/GLHelpers.cpp
@@ -2,20 +2,16 @@
 
 #include <mutex>
 
+#include <QtCore/QObject>
 #include <QtCore/QThread>
 #include <QtCore/QRegularExpression>
 #include <QtCore/QProcessEnvironment>
+
 #include <QtGui/QSurfaceFormat>
+#include <QtGui/QOpenGLContext>
+#include <QtGui/QOpenGLDebugLogger>
+
 #include <QtOpenGL/QGL>
-#include <QOpenGLContext>
-
-#ifdef DEBUG
-static bool enableDebug = true;
-#else
-static const QString DEBUG_FLAG("HIFI_ENABLE_OPENGL_45");
-static bool enableDebug = QProcessEnvironment::systemEnvironment().contains(DEBUG_FLAG);
-#endif
-
 
 const QSurfaceFormat& getDefaultOpenGLSurfaceFormat() {
     static QSurfaceFormat format;
@@ -25,7 +21,8 @@ const QSurfaceFormat& getDefaultOpenGLSurfaceFormat() {
         format.setDepthBufferSize(DEFAULT_GL_DEPTH_BUFFER_BITS);
         format.setStencilBufferSize(DEFAULT_GL_STENCIL_BUFFER_BITS);
         setGLFormatVersion(format);
-        if (enableDebug) {
+        if (GLDebug::enabled()) {
+            qDebug() << "Enabling debug context";
             format.setOption(QSurfaceFormat::DebugContext);
         }
         format.setProfile(QSurfaceFormat::OpenGLContextProfile::CoreProfile);
@@ -81,3 +78,29 @@ bool isRenderThread() {
     return QThread::currentThread() == RENDER_THREAD;
 }
 
+
+#ifdef DEBUG
+static bool enableDebugLogger = true;
+#else
+static const QString DEBUG_FLAG("HIFI_DEBUG_OPENGL");
+static bool enableDebugLogger = QProcessEnvironment::systemEnvironment().contains(DEBUG_FLAG);
+#endif
+
+bool GLDebug::enabled() {
+    return enableDebugLogger;
+}
+
+void GLDebug::log(const QOpenGLDebugMessage & debugMessage) {
+    qDebug() << debugMessage;
+}
+
+void GLDebug::setupLogger(QObject* window) {
+    if (enabled()) {
+        QOpenGLDebugLogger* logger = new QOpenGLDebugLogger(window);
+        logger->initialize(); // initializes in the current context, i.e. ctx
+        logger->enableMessages();
+        QObject::connect(logger, &QOpenGLDebugLogger::messageLogged, window, [&](const QOpenGLDebugMessage & debugMessage) {
+            GLDebug::log(debugMessage);
+        });
+    }
+}
